Revert cog system changes
