+++ discord/__init__.py
@@ -61,7 +61,6 @@ from .components import *
 from .threads import *
 from .bot import *
 from .app import *
-from .cog import Cog
 from .welcome_screen import *
