+++ discord/ext/commands/core.py
@@ -48,10 +48,9 @@ import datetime
 import discord
 
 from .errors import *
-from ...errors import *
 from .cooldowns import Cooldown, BucketType, CooldownMapping, MaxConcurrency, DynamicCooldownMapping
 from .converter import run_converters, get_converter, Greedy
-from ...app import _BaseCommand, slash_command, user_command, message_command
+from ._types import _BaseCommand
 from .cog import Cog
 from .context import Context
 
@@ -94,10 +93,7 @@ __all__ = (
     'is_owner',
     'is_nsfw',
     'has_guild_permissions',
-    'bot_has_guild_permissions',
-    'slash_command',
-    'user_command',
-    'message_command'
+    'bot_has_guild_permissions'
 )
 
 MISSING: Any = discord.utils.MISSING
